+++ src/main/java/Datos/Evento/EventoRifaFactory.java
@@ -4,17 +4,17 @@ import java.util.ArrayList;
 
 public class EventoRifaFactory extends EventoFactory{
     @Override
-    public iEvento createEvento(String nombre, String ubicacion, int objetivoRecaudacion, String descripcion, String date, String url, String informacion) {
+    public iEvento createEvento(String nombre, String ubicacion, float objetivoRecaudacion, String descripcion, String date, String url, String informacion) {
         return new EventoRifa(nombre, ubicacion, objetivoRecaudacion, descripcion, date, url);
     }
 
     @Override
-    public iEvento createEvento(String nombre, String ubicacion, int recaudacion, int objetivoRecaudacion, String descripcion, String date, String url, String informacion) {
+    public iEvento createEvento(String nombre, String ubicacion, float recaudacion, float objetivoRecaudacion, String descripcion, String date, String url, String informacion) {
         return new EventoRifa(nombre, ubicacion, recaudacion, objetivoRecaudacion, descripcion, date, url);
     }
 
     @Override
-    public iEvento createEvento(String nombre, String ubicacion, int recaudacion, int objetivoRecaudacion, String descripcion, String date, String url, int ID, String informacion) {
+    public iEvento createEvento(String nombre, String ubicacion, float recaudacion, float objetivoRecaudacion, String descripcion, String date, String url, int ID, String informacion) {
         return new EventoRifa(nombre, ubicacion, recaudacion, objetivoRecaudacion, descripcion, date, url, ID);
     }
 }
